Validate slot format rather than existence
Previously the validation would call the get_slots API again in order to validate that the picked slot was available (on submitting both the slots step and the confirmation step.

This resulted in a lot of unnecessary calls to the get_slots API.

Also, as the slots returned by get_slots becomes more dynamic, it feels wrong to re-validate an earlier step which could change later due to a slot becoming unavailable.

I have therefore amended to simply validate the format. The call to request_visit should fail if the requested slot is not available for booking.