Refactor steps processor (pass processor to steps)
This allows steps to refer back to the processor for information related to the entire booking flow, rather than passing things like the prison id to each step.

This will be particularly relevant as additional prisoner information is required in order to present the slots step.

Note: temporarily disabled limiting of visitors – this validation is in the wrong place