Remove duplicate locale in request visit
We pass the locale in the Accept-Language so having the locale parameter when
requesting a booking is redundant.

As part of this change I noticed the VCR wasn't strict enough when matching the
requests. The default configuration only matches on the http method and URI.
This meant that changing the body or headers matched a previously recorded
cassette with different values. I changed the default configuration for our
usage to match on all possible values including path, body and headers.

Had to manually change the cassettes as I don't have access to the NOMIS api
away from the office. We should regenerate the cassettes when possible.