Validate visit processing_state received from API
This is important because we are rendering templates dependant on the processing state, and we don’t want to get an exception warning of a missing template.

This required recording new VCR specs – the API has moved on and now returns processing_state from all visit calls.

Removed the unnecessary cy locale which complicated matters with unrelated failures.

Brakeman was warning against this render call, but it’s now safe to ignore (it’s now a false-positive).