Group instruments and make explictly threadsafe
The `append_to_log` method is an instrument, and so does not really
belong in the applcation controller.  This felt like a better home for
it.  The instance variable would have also been threadsafe in that a
request is going to happen in a single thread, but RequestStore feels
like a more declarative/explicit solution under the circumstances.