Handle invalid query strings while error handling
Our exception app is an ActionController action (they are Rack apps), the
problem is that if an error happens inside that application it will be bubble
out of the error app resulting in a non-styled error page showed to the user and
the error not being reported to Sentry.

We've seen this happen with invaid query strings that involve invalid nested
attributes. To fix this we need to detect if the query string is invalid before
is processed by the exception app, otherwise the controller will try to parse
the query string before processing the main body of the action and raise an
error.

An alternative of doing this would be of decoupling our error pages from Rails
controllers / views, and have a Rack app that has the same behaviour while not
having to use the query string.