Add text to requested & booked pages, and DRY views
The text on the requested page is now closer to what was previously on the status page for requested bookings (rather than the confirmation page which had inconsistent messaging r.e. what to do if emails didn’t arrive).

The booked page text is new, and draws attention to the key actions.

In both views, we link to the cancellation page which is now less obvious at the bottom of the page.

Duplicated visit info removed into partials.