Fix JS race conditon in prison autocomplete
We were getting random failures because the prison autocomplete hadn't yet
loaded when we were filling the prison field. This meant that the form was
sending a nil `prison_id`.

The fix is to click on the element from the autocomplete library to ensure that
the JS functionality has loaded and is ready.