Fix email validation bug
Whilst there is an MX check that is included in the EmailChecker for
feedback submission it did not raise any errors when there was an
invalid format e.g. "user@example..com".

The first step of the fix was to amend the HTML input type to "email"
In newer browsers this will pick up/prompt the user where an invalid
email format has been added and will prevent submission until corrected.

However, this fix does not support older browsers and so the
 #domain_dot_error method has been extended to check for two or more
 consecutive dots in the domain, and raise an error if found.