Downgrade puma to 3.9.1
Puma got bumped around the same time we started having production issues.

I have doubts it's the cause of the issue since pvb2 runs the same version and
there are no reported issues.

We can bump the version again if the issue happens again or after a period of no
issues to see if it reoccurs.