Retry socket errors once for POST requests
If the socket gets closed and excon tries to use it a socket error is be raised.

When this happens for idempotent requests Excon retries automatically
restablishing the connection. The problem is for POST requests, these are not
retried so the user ends up with an error page.

This introduces the risk of double processing POST requests so we should keep an
eye on this and revisit if necessary.