Upgrade to Rails 5.2 and remove Virtus
Remove Virtus and use the new attributes api in ActiveModel that supports types.

The main difference between Virtus and ActiveModel is that Virtus provides a
coercer which can return different types, while ActiveModel approach is to
specify what type the attribute should reference via a coercer that must return
that type. This meant that we needed a different approach for handling invalid
dates than the solution implemented for Virtus via MaybeDate.

Another key difference is that ActiveModel does not support typed arrays. In
Virtus one can specify an attribute to be for example `Array[Visitor]`.

There are other cleanup changes in this PR such as removing attributes that are
no longer in use in the API and consolidating the 2 Visitor classes into one.