Remove active-model-attributes gem & upgrade Rail 5.2
We are now in a position to upgrade to Rails 5.2 and with this we need to
remove the 'active-model-attributes' gem and use of ActiveModel::Attributes
instead.

In addition to this there are some attribute type lists are being populated
with unnecessary types and do not provide much value. Removing these
unnecessary types will make the code easier to read and reduce complexity.