Prevent search engines indexing site
The service manual says that all services must tell search engines not
to index pages on their domain so that users start from GOV.UK rather
going directly to the service domain.  We have not been doing this and so
the public service shows up in Google results.  This PR implements the
X-Robots-Tag header, which applies to all content types in the app (e.g.
images, scripts, styles), not only HTML files.

Related to this Trello card:
https://trello.com/c/uTzGYsBk/15-prevent-search-engines-indexing-the-public-site