Add flash notice re: service migration
We are due to migrate PVB from template deploy to kubernetes by the end
of the week. Our intention is to create a maintenance page, but we would
like to forewarn the public by displaying a flash notice for the next
couple of days.  This flash notice is displayed on the 'Who are you
visiting' & 'When would you like to visit' pages only.

This flash alert PR will need to be reverted as part of the migration
tasks to ensure the public do not see it once the work has been completed.