Make pods safely moveable
Cloudplatform are making changes to ensure that pods don't get too old,
and to do this they will be randomly killing older pods from time to
time (and/or migrating them to other nodes).

This PR makes the recommended changes to ensure the system can safely
stay up and running during this process.

The number of pods is also increased (inline with the recommendation)
and so it is probably best to ensure the resources used by the new pod
are not excessive.