POM-431 - Reduce CircleCI workflow steps
We currently have three separate workflow steps in CircleCI for running
tests, security analysis and Rubocop. This means that lots of jobs are
created within CirceCI each time we push a branch/want to deploy etc and
so this PR puts them into one step, which should make the CI/CD process
a bit more efficient.