Fix the Welsh calendar journey
Whilst working on another piece of work I discovered that the calendar /
slot picker step within the Welsh flow was broken due to an issue with
the formatting of the day in the yaml file.

In addition to this I have discovered a number of missing translations
that were also impacting the service and have added these.