POM 600 - Fix PVB content
We’ve had several requests for the service to clarify to users that they
are sending a request for a booking, and not a guarantee. Although the
email clarifies this it was agreed that we needed to update the content
on the site to be more explicit.  This PR updates the content in both
English and Welsh.